uimac: Type fixes to silence compiler warnings
NSInteger and NSUInteger are Apple types from Cocoa. They are defined
differently depending on the target architectures. On a 64-bit platform
they are 64-bit signed and unsigned integer, and they are 32-bit signed
and unsigned integer on 32-bit platforms.

Thus, we cast to long and use %ld (or unsigned long and %lu)--as is
commonly done with these types--assuming that NSInteger will always fit
in long. (We could cast to int64_t and use PRId64, assuming NSInteger
never becomes bigger than int64_t.)

+++ src/uimac/ProfileController.m
@@ -18,8 +18,8 @@ NSString *unisonDirectory()
 #else
     NSArray *files = [[NSFileManager defaultManager] contentsOfDirectoryAtPath:directory error:nil];
 #endif
-    unsigned int count = [files count];
-    unsigned int i,j;
+    NSUInteger count = [files count];
+    NSUInteger i,j;
 
     [profiles release];
     profiles = [[NSMutableArray alloc] init];
@@ -55,7 +55,7 @@ NSString *unisonDirectory()
     [tableView reloadData];
 }
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSUInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     if (!profiles) return 0;
     else return [profiles count];
@@ -72,7 +72,7 @@ NSString *unisonDirectory()
 
 - (NSString *)selected
 {
-    int rowIndex = [tableView selectedRow];
+    NSInteger rowIndex = [tableView selectedRow];
     if (rowIndex >= 0 && rowIndex < [profiles count])
         return [profiles objectAtIndex:rowIndex];
     else return nil;
